Move useTimeout hook to Canonical
